Updates to dateTime and Time Zones
Looks like this does not exist: Get datetime ODM2 CV for column name:

Add check mark for DST observed or no - if yes, subset the list for DST options, if no, subset for list

Attached is a .csv and .txt (tab delim) file with indication of whether or not DST is observed for each time zone. User will indicate whether or not the dataset observes DST, and that will limit the time zone list to those that match DST/no DST.

Add to top of ezPrepare: "For initial development, this tool was designed to prepare underwater ecological data for upload to the EDI data portal."

When a user types in their Date Time format, check if a valid Date Time object can be made using Luxon. Display a message stating if it is valid or not. Display the message in the help text section

Add hover with TimeZone selection: “For more information on timezone names and details, see this link.”